requirements.txt
pytest
requests
python-dotenv
voluptuous
pytest_voluptuous
curlify
allure-pytest